Rename source modules to executable names
as having them named the same is customary for go.

Also build the example in c that uses the library.

Use the implicit output names in the Makefile.

Change the README to refer to the Makefile instead of duplicating its
content.

Move all output in a subdirectory named out, so that source and binary
names don't conflict.

Move examples that use libsuseconnect.so to libsuseconnect-examples.
