+++ .gitignore
@@ -1,4 +1,1 @@
-suseconnect
-libsuseconnect.h
-libsuseconnect.so
-use-lib
+out
