+++ Makefile
@@ -1,16 +1,22 @@
 all: test build build-so
 
-build:
-	go build cmd/suseconnect.go
+out:
+	mkdir -p out
+
+build: out
+	go build -v -o out/ github.com/SUSE/connect-ng/suseconnect
 
 test:
 	go test -v ./connect
 
-build-so:
-	go build -buildmode=c-shared -o libsuseconnect.so ext/main.go
+build-so: out
+	go build -v -buildmode=c-shared -o out/libsuseconnect.so github.com/SUSE/connect-ng/libsuseconnect
 
-build-arm:
-	GOOS=linux GOARCH=arm64 GOARM=7 go build cmd/suseconnect.go
+build-so-example: build-so
+	gcc libsuseconnect-examples/use-lib.c -o out/use-lib -Lout -lsuseconnect
+
+build-arm: out
+	GOOS=linux GOARCH=arm64 GOARM=7 go build -v -o out/ github.com/SUSE/connect-ng/suseconnect
 
 clean:
 	go clean
