+++ libsuseconnect-examples/use-lib.c
@@ -1,12 +1,11 @@
 /*
-    go build -buildmode=c-shared -o libsuseconnect.so ext/main.go
-    gcc ext/use-lib.c -o use-lib -L. -lsuseconnect
-    LD_LIBRARY_PATH=. ./use-lib
+    cd ..
+    make build-so-example
 */
 
 #include <stdio.h>
 #include <stdlib.h>
-#include "../libsuseconnect.h"
+#include "../out/libsuseconnect.h"
 
 void main(void) {
     char *format = "json";
