+++ libsuseconnect-examples/use-lib.rb
@@ -12,7 +12,7 @@ end
 
 module SUSEConnect
   extend FFI::Library
-  ffi_lib './libsuseconnect.so'
+  ffi_lib './out/libsuseconnect.so'
 
   attach_function :getstatus, [:string], :pointer
 end
