+++ libsuseconnect/libsuseconnect.go
