+++ suseconnect/suseconnect.go
